Modify TextBlock to handle multiple fonts better
Added logic to allow for the text height to be measured by the line
rather than by the last item. It seems to work really well. Next step
is to make TextBlock.cs have an interface and make DialogBox an abstract
class, so that more work can be done to make everything modular.